requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
cycler==0.10.0
dateparser==1.0.0
idna==3.2
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
multidict==5.1.0
multitasking==0.0.9
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
pyparsing==2.4.7
python-binance==1.0.12
python-dateutil==2.8.2
python-kucoin==2.1.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
TA-Lib==0.4.21
typing-extensions==3.10.0.1
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websockets==9.1
yarl==1.6.3
yfinance==0.1.63
